requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
croniter==6.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2024.12.0
grandalf==0.8
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.78
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-mongodb==0.7.0
langchain-neo4j==0.5.0
langchain-openai==0.3.33
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
langgraph==1.0.0a4
langgraph-checkpoint==2.1.1
langgraph-checkpoint-mongodb==0.2.1
langgraph-prebuilt==0.7.0a2
langgraph-sdk==0.2.9
langsmith==0.4.31
lark==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mem0ai==0.1.118
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.10.0
networkx==3.5
numpy==2.3.3
openai==1.108.2
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
portalocker==3.2.0
posthog==6.7.6
propcache==0.3.2
protobuf==5.29.5
PyAudio==0.2.14
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.15.2
pyparsing==3.2.5
pypdf==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
rank-bm25==0.2.2
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rq==2.6.0
safetensors==0.6.2
scipy==1.16.2
sentry-sdk==2.38.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.8
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
types-PyYAML==6.0.12.20250915
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
